Amend docker setup for local fake data

+++ script/fake_data.py
@@ -20,12 +20,26 @@ from nowcasting_datamodel.models.forecast import (
     ForecastValueSQL,
 )
 from nowcasting_datamodel.models.models import StatusSQL
+from sqlalchemy import inspect
 
 from src.utils import floor_30_minutes_dt
 
 now = floor_30_minutes_dt(datetime.now(tz=timezone.utc))
 
+if os.getenv("DB_URL") is None:
+    raise ValueError("Please set the DB_URL environment variable")
+
+if not os.getenv("DB_URL") == "postgresql://postgres:postgres@postgres:5432/postgres":
+    raise ValueError("This script should only be run in the local docker container")
+
 connection = DatabaseConnection(url=os.getenv("DB_URL", "not_set"))
+
+print(f"Creating fake data at {now}")
+
+# Check database has been created, if not run the migrations
+if not inspect(connection.engine).has_table("status"):
+    connection.create_all()
+
 
 with connection.get_session() as session:
     session.query(StatusSQL).delete()
@@ -48,7 +62,7 @@ with connection.get_session() as session:
     make_fake_gsp_yields(gsp_ids=range(0, N_GSPS), session=session, t0_datetime_utc=now)
 
     # 3. make status
-    status = StatusSQL(status="warning", message="this is all fake data")
+    status = StatusSQL(status="warning", message="Local Quartz API serving fake data")
 
     session.add(status)
     session.commit()
